refactor(pipeline): move config and prefix to pipeline/schemas/
Aligns with the directory layout in docs/design/data-pipeline.md §14
which specifies pipeline/schemas/ for Pydantic models and ID helpers.

Refs #267
